requirements.txt
flask_sqlalchemy
psycopg2-binary
flask
waitress
python-dotenv